move `groupBySingle` into `array-utils`
